requirements.txt
python-telegram-bot[job-queue]>=20.3,<21
requests>=2.32
tzdata>=2025.1
python-dotenv>=1.0